Partial upgrade to signposts
Started building system to handle diplomacy when player interacts with signs. It's not working right yet, in part because 'compass' is always going to be 3, rather than vary from 0 to 3 depending on which sign is visited. Also I need to actually write the code for "do envoy stuff".

In addition, this commit contains miscellaneous dialog fixes, e.g., more conversion of sixth-days to work-shifts, and a more cheerful final victory message.

Finally, iron production was added to autumn and winter, since it seems to be behaving well.